+++ changelog.md
@@ -14,9 +14,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.0.4] - After fiestas-2022
 
+### Update
+
 - `arch` Update Angular Core from 14 to 15.
 - `arch` Update Angular CDK + Material from 14 to 15.
 - `style` Fix Styles after Angular Material 15.
+
+### Delete
+
+- Remove temporally angular universal.
+
 
 ## [2.0.3] - After fiestas-2022